backend/modules/veo3_integration/generator.py
"""
Google VEO3 Integration Module
Handles video generation using Google's VEO3 model
"""

import os
import asyncio
import logging
import json
from typing import Dict, Any, Optional, List
from enum import Enum
from dataclasses import dataclass
from pathlib import Path

import httpx
import vertexai
from vertexai.generative_models import GenerativeModel
from google.cloud import aiplatform

logger = logging.getLogger(__name__)

class VEO3Quality(Enum):
    """VEO3 Quality settings"""
    VEO_3 = "veo-3"
    VEO_3_HD = "veo-3-hd"

class VEO3AudioMode(Enum):
    """VEO3 Audio modes"""
    FULL = "full"
    MUSIC_ONLY = "music_only"
    SILENT = "silent"

class VEO3Resolution(Enum):
    """VEO3 Resolution options"""
    RES_720P = "720p"
    RES_1080P = "1080p"
    RES_4K = "4k"

@dataclass
class VEO3Config:
    """VEO3 Configuration"""
    quality: VEO3Quality = VEO3Quality.VEO_3
    duration: int = 8  # Max 8 seconds
    audio_mode: VEO3AudioMode = VEO3AudioMode.FULL
    resolution: VEO3Resolution = VEO3Resolution.RES_720P
    temperature: float = 0.7
    seed: Optional[int] = None

@dataclass
class VEO3GenerationRequest:
    """VEO3 Generation Request"""
    prompt: str
    config: VEO3Config
    reference_image: Optional[str] = None
    style_reference: Optional[str] = None

@dataclass
class VEO3GenerationResult:
    """VEO3 Generation Result"""
    video_url: str
    duration: float
    resolution: str
    file_size: int
    generation_time: float
    metadata: Dict[str, Any]

class GoogleVEO3Generator:
    """Google VEO3 Video Generator"""
    
    def __init__(self):
        self.project_id = os.getenv('GOOGLE_CLOUD_PROJECT')
        self.location = os.getenv('GOOGLE_CLOUD_LOCATION', 'us-central1')
        self.api_key = os.getenv('GOOGLE_API_KEY')
        
        # Initialize Vertex AI
        if self.project_id:
            try:
                vertexai.init(project=self.project_id, location=self.location)
                logger.info("Vertex AI initialized successfully")
            except Exception as e:
                logger.warning(f"Failed to initialize Vertex AI: {e}")
        else:
            logger.warning("Google Cloud project not configured")
    
    async def generate_video(self, request: VEO3GenerationRequest) -> VEO3GenerationResult:
        """Generate video using VEO3"""
        try:
            logger.info(f"Starting VEO3 video generation with prompt: {request.prompt[:100]}...")
            
            # For now, create a mock implementation
            # In production, this would use the actual VEO3 API
            result = await self._generate_mock_video(request)
            
            logger.info("VEO3 video generation completed successfully")
            return result
            
        except Exception as e:
            logger.error(f"VEO3 generation failed: {str(e)}")
            raise
    
    async def _generate_mock_video(self, request: VEO3GenerationRequest) -> VEO3GenerationResult:
        """Mock video generation for testing"""
        import time
        import random
        
        # Simulate generation time
        generation_start = time.time()
        await asyncio.sleep(5)  # Simulate 5 second generation
        generation_time = time.time() - generation_start
        
        # Create mock result
        video_filename = f"veo3_video_{int(time.time())}_{random.randint(1000, 9999)}.mp4"
        video_path = f"assets/clips/{video_filename}"
        
        # Create mock video file (empty for now)
        os.makedirs(os.path.dirname(video_path), exist_ok=True)
        with open(video_path, 'w') as f:
            f.write("Mock VEO3 video file")
        
        return VEO3GenerationResult(
            video_url=f"/assets/clips/{video_filename}",
            duration=request.config.duration,
            resolution=request.config.resolution.value,
            file_size=1024 * 1024,  # 1MB mock
            generation_time=generation_time,
            metadata={
                "prompt": request.prompt,
                "config": {
                    "quality": request.config.quality.value,
                    "audio_mode": request.config.audio_mode.value,
                    "temperature": request.config.temperature,
                    "seed": request.config.seed
                },
                "model": "veo-3",
                "generated_at": time.time()
            }
        )
    
    async def _call_veo3_api(self, request: VEO3GenerationRequest) -> Dict[str, Any]:
        """Call the actual VEO3 API (placeholder)"""
        # This would implement the actual VEO3 API call
        # Using Vertex AI or direct API calls
        
        api_request = {
            "prompt": request.prompt,
            "config": {
                "quality": request.config.quality.value,
                "duration": request.config.duration,
                "audio_mode": request.config.audio_mode.value,
                "resolution": request.config.resolution.value,
                "temperature": request.config.temperature
            }
        }
        
        if request.reference_image:
            api_request["reference_image"] = request.reference_image
        
        if request.config.seed:
            api_request["seed"] = request.config.seed
        
        # Mock API response
        return {
            "video_url": "https://mock-veo3-api.com/video.mp4",
            "duration": request.config.duration,
            "status": "completed"
        }
    
    def get_available_models(self) -> List[str]:
        """Get list of available VEO3 models"""
        return [
            "veo-3",
            "veo-3-hd"
        ]
    
    def validate_prompt(self, prompt: str) -> bool:
        """Validate prompt for VEO3 generation"""
        if not prompt or len(prompt.strip()) == 0:
            return False
        
        if len(prompt) > 1000:  # Arbitrary limit
            return False
        
        return True
    
    def estimate_generation_time(self, config: VEO3Config) -> float:
        """Estimate generation time based on config"""
        base_time = 30  # Base 30 seconds
        
        # Adjust for quality
        if config.quality == VEO3Quality.VEO_3_HD:
            base_time *= 2
        
        # Adjust for duration
        base_time *= (config.duration / 4)  # Scale with duration
        
        # Adjust for resolution
        if config.resolution == VEO3Resolution.RES_1080P:
            base_time *= 1.5
        elif config.resolution == VEO3Resolution.RES_4K:
            base_time *= 3
        
        return base_time

# Utility functions
def create_veo3_config(
    quality: str = "veo-3",
    duration: int = 8,
    audio_mode: str = "full",
    resolution: str = "720p",
    temperature: float = 0.7,
    seed: Optional[int] = None
) -> VEO3Config:
    """Create VEO3 configuration from parameters"""
    return VEO3Config(
        quality=VEO3Quality(quality),
        duration=min(duration, 8),  # Max 8 seconds
        audio_mode=VEO3AudioMode(audio_mode),
        resolution=VEO3Resolution(resolution),
        temperature=max(0.0, min(1.0, temperature)),
        seed=seed
    )

def create_generation_request(
    prompt: str,
    quality: str = "veo-3",
    duration: int = 8,
    audio_mode: str = "full",
    resolution: str = "720p",
    temperature: float = 0.7,
    seed: Optional[int] = None,
    reference_image: Optional[str] = None,
    style_reference: Optional[str] = None
) -> VEO3GenerationRequest:
    """Create VEO3 generation request"""
    config = create_veo3_config(
        quality=quality,
        duration=duration,
        audio_mode=audio_mode,
        resolution=resolution,
        temperature=temperature,
        seed=seed
    )
    
    return VEO3GenerationRequest(
        prompt=prompt,
        config=config,
        reference_image=reference_image,
        style_reference=style_reference
    )